Backend/requirements.txt
Flask==2.0.2
numpy==1.25.2
opencv_python==4.7.0.68
Pillow==9.1.0
# Pillow==10.0.0
tensorflow==2.12.0
tensorflow_intel==2.12.0
Werkzeug==2.0.2